Collect `currentFile` context data separate from workfile instance
